LC-powx-n [Runtime: N/A; Memory: N/A]

+++ math/powx-n.cpp
@@ -7,7 +7,7 @@ public:
         double ans=1;
 
         if (n>0){
-            for(int i=1; i<=n; i++){
+            for(double i=1; i<=n; i++){
                 ans*=x;
             }
         }
